🧪 test scores.service.ts

+++ src/scores/scores.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { ScoreType } from '@prisma/client';
-import { PrismaService } from 'src/prismaService/prisma.service';
+import { PrismaService } from 'src/prisma/prisma.service';
 import { CoreOutput } from '../common/dtos/output.dto';
 import { CreateScoreInput, ScoreOutput } from './dto/create-score.input';
 import { ScoresOutput } from './dto/scores.dto';
@@ -19,12 +19,15 @@ export class ScoresService {
     username,
   }: CreateScoreInput): Promise<ScoreOutput> {
     try {
-      const userExists = await this.prisma.user.findUnique({
+      const userExist = await this.prisma.user.findUnique({
         where: { username },
       });
 
-      if (!userExists) {
-        throw new Error('존재하지 않는 학생입니다.');
+      if (!userExist) {
+        return {
+          success: false,
+          error: '존재하지 않는 학생입니다.',
+        };
       }
 
       const createdScore = await this.prisma.score.create({
@@ -41,7 +44,7 @@ export class ScoresService {
 
       return { success: true, score: createdScore };
     } catch (e) {
-      return { success: false, error: e.message };
+      return { success: false, error: "Can't create score" };
     }
   }
 
@@ -78,18 +81,27 @@ export class ScoresService {
     } catch (e) {
       return {
         success: false,
-        error: e.message,
+        error: "Can't create score",
       };
     }
   }
 
   async remove(id: string): Promise<ScoreOutput> {
     try {
+      const scoreExist = await this.prisma.score.findUnique({ where: { id } });
+
+      if (!scoreExist) {
+        return {
+          success: false,
+          error: '존재하지 않는 점수입니다',
+        };
+      }
+
       const score = await this.prisma.score.delete({ where: { id } });
 
       return { success: true, score };
     } catch (e) {
-      return { success: false, error: e.message };
+      return { success: false, error: "Can't remove score" };
     }
   }
 
@@ -105,7 +117,7 @@ export class ScoresService {
         scores,
       };
     } catch (e) {
-      return { success: false, error: e.message };
+      return { success: false, error: "Can't find scores" };
     }
   }
 
@@ -115,7 +127,7 @@ export class ScoresService {
 
       return { success: true };
     } catch (e) {
-      return { success: false, error: e.message };
+      return { success: false, error: "Can't reset the score" };
     }
   }
 }
